Replace old factory method
